Made modification using String.format()

+++ src/dir/shape.java
@@ -3,5 +3,5 @@ package dir;
 public  interface shape {
     public void getArea();
     public void getPerimeter();
-    public void setside(double side);
+    public void setside(float side);
 }
